requirements.txt
# Python version requirement: 3.10.11

# Core dependencies
python-dotenv>=1.0.0
neo4j>=5.0.0
llama-index>=0.9.0
llama-index-graph-stores-neo4j>=0.2.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
openai>=1.0.0
numpy>=1.24.0

# Phoenix and OpenTelemetry dependencies for tracing
arize-phoenix>=4.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
openinference-instrumentation-llama-index>=2.0.0
openinference-semantic-conventions>=0.1.0

# Streamlit for web interface
streamlit>=1.28.0

# Additional dependencies that may be required by the above packages
typing-extensions>=4.0.0
pathlib2>=2.3.0  # For older Python versions compatibility
asyncio>=3.4.4  # Usually built-in, but listed for completeness
